requirements_dev.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bidict==0.21.2
biosppy==0.6.1
bleach==3.2.1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
coverage==5.3.1
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
ecq-qc==1.0b1
entrypoints==0.3
flake8==3.8.4
gatspy==0.3
h5py==3.1.0
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
keyring==21.7.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mne==0.22.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nose==1.3.7
notebook==6.1.5
numpy==1.19.4
packaging==20.8
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkginfo==1.6.1
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.10.0
py-ecg-detectors==1.0.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
SecretStorage==3.3.0
Send2Trash==1.5.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
twine==3.2.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
wfdb==3.1.1
widgetsnbextension==3.5.1
xgboost==1.3.0.post0
zipp==3.4.0
